Fix Sprint 3G
Removed bearer for Sprint eHRPD and LTE APNs. The eHRPD APN was set
to bearer 13 even though it is valid for all 1x/EVDO/eHRPD connecti
ons. This was causing no APN to be selected when an EVDO connection
was established and radioTech != 13 (eg EVDO_REV_A = 8). The LTE be
arer must also be unspecified so telephony does not then select the
eHRPD APN since any unspecified bearer is potentially valid for bot
h 3G and LTE. Regardless, the bearer does not need to be specified
since the system selects the correct APN for 3G or LTE automaticall
y.

Change-Id: I4ab61625888bf793a774fca2baab8274760452c4
Signed-off-by: AudioGod <AudioGod@Sonic-Developers.com>